refactor: 将 command 模块移动到 core 目录 & 更新导入
将 command 相关文件从 `src/command` 移动到 `src/core/command`

更新相关文件的导入路径
