+++ src/core/command/command.ts
