+++ src/core/command/errors.ts
