+++ src/core/command/manager.ts
