+++ src/plugins/chatSpawn.ts
@@ -8,7 +8,7 @@ import {
     spawnSimulatedPlayer,
     spawnSimulatedPlayerByNameTag
 } from './main'
-import { type CommandInfo, commandManager, Command } from '../command'
+import { type CommandInfo, commandManager, Command } from '../core/command'
 import { Dimension, Vector3, world, type Player } from '@minecraft/server'
 import {xyz_dododo} from "../utils/xyz_dododo";
